Add debug-print script to jstests package.json
