+++ composer/modules/js-tests/package.json
@@ -1,8 +1,8 @@
 {
   "scripts": {
     "test": "mocha --timeout 3000 src/test/js/test.js",
-    "test:debug": "cross-env SOURCE_DEBUG=true mocha target/test/js/source-gen-test.js",
-    "compile": "babel ./src --out-dir ./target"
+    "debug-print": "cross-env DEBUG=true mocha --timeout 3000 src/test/js/test.js --grep \"debug print\"",
+    "test:debug": "cross-env SOURCE_DEBUG=true mocha target/test/js/source-gen-test.js"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
