+++ composer/modules/js-tests/pom.xml
@@ -167,21 +167,6 @@
                         </configuration>
                     </execution>
                     <execution>
-                        <id>npm compile (run babel)</id>
-                        <goals>
-                            <goal>exec</goal>
-                        </goals>
-                        <phase>compile</phase>
-                        <configuration>
-                            <executable>${npm.executable}</executable>
-                            <arguments>
-                                <argument>run</argument>
-                                <argument>compile</argument>
-                            </arguments>
-                            <skip>${maven.test.skip}</skip>
-                        </configuration>
-                    </execution>
-                    <execution>
                         <id>npm test</id>
                         <goals>
                             <goal>exec</goal>
