+++ composer/modules/js-tests/src/test/js/test.js
@@ -18,17 +18,16 @@
 
 const fs = require('fs');
 const path = require('path');
-const request = require('request');
 const { expect } = require('chai');
 const { spawn } = require('child_process');
 
 global.targetPath = path.join(__dirname, '..', '..', '..', 'target');
 
 const testEnv = require('./environment');
+const {findBalFilesInDirSync, parse} = require('./utils');
+const debugPrint = require('./source-gen-debug-print');
 
 const testFilesDir = path.join(__dirname, '../resources/ballerina-examples');
-const parserUrl = `http://127.0.0.1:9091/composer/ballerina/parser/file/validate-and-parse`;
-
 
 describe('Ballerina Composer Test Suite', () => {
     const testFiles = findBalFilesInDirSync(testFilesDir);
@@ -69,51 +68,27 @@ describe('Ballerina Composer Test Suite', () => {
                         done();
                     });
                 })
-            })
+            });
 
             it('renders', () => {
                 testEnv.render(model)
-            })
+            });
 
             it('generates source', () => {
                 const generatedSource = testEnv.generateSource(model);
                 expect(generatedSource).to.equal(content);
-            })
-        })
-    })
+            });
+
+            if (process.env.DEBUG == "true") {
+                it('debug print', () => {
+                    const tree = testEnv.buildTree(model);
+                    debugPrint(tree);
+                });
+            }
+        });
+    });
 
     after(() => {
         backEndProcess.kill();
     });
 });
-
-function findBalFilesInDirSync(dir, filelist) {
-    const files = fs.readdirSync(dir);
-    filelist = filelist || [];
-    files.forEach((file) => {
-        if (fs.statSync(path.join(dir, file)).isDirectory()) {
-            filelist = findBalFilesInDirSync(path.join(dir, file), filelist);
-        } else if (path.extname(file) === '.bal') {
-            filelist.push(path.join(dir, file));
-        }
-    });
-    return filelist;
-}
-
-function parse(content, file, callback) {
-    const parseOpts = {
-        content,
-        fileName: path.basename(file),
-        filePath: path.dirname(file),
-        includePackageInfo: true,
-        includeProgramDir: true,
-        includeTree: true,
-    }
-
-    const parseReq = request.post({
-        url: parserUrl,
-        json: parseOpts,
-    }, (err, httpResponse, body) => {
-        callback(body.model)
-    });
-}
